+++ src/components/Post.tsx
@@ -7,6 +7,7 @@ import Card1 from "./cards/Card1";
 import Card2 from "./cards/Card2";
 import Row1 from "./cards/Row1";
 import CommentCard from "./cards/CommentCard";
+import { useRead } from "../hooks/useRead";
 
 const Post = ({ post, postNum = 0 }) => {
   const context: any = useMainContext();
@@ -17,6 +18,8 @@ const Post = ({ post, postNum = 0 }) => {
   const { data: session, status } = useSession();
   const [hasMedia, setHasMedia] = useState(false);
   const [margin, setMargin] = useState("m-1");
+  const {read} = useRead(post?.data?.name)
+  const [commentsDirect, setCommentsDirect] = useState(false); 
 
   useEffect(() => {
     context.nsfw === false && post?.data?.over_18
@@ -47,16 +50,20 @@ const Post = ({ post, postNum = 0 }) => {
     if (lastRoute === router.asPath) {
       //console.log("match");
       setSelect(false);
+      setCommentsDirect(false);
       context.setPauseAll(false);
     }
     //don't add lastRoute to the array, breaks things
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [router.asPath]);
 
-  const handleClick = (e) => {
+  const handleClick = (e, toComments) => {
     e.stopPropagation();
     // plausible("postOpen");
-    if (!e.ctrlKey) {
+    if (toComments){
+      setCommentsDirect(true); 
+    }
+    if (!e.ctrlKey && !e.metaKey) {
       setLastRoute(router.asPath);
       context.setPauseAll(true);
       setSelect(true);
@@ -132,15 +139,14 @@ const Post = ({ post, postNum = 0 }) => {
           postData={post?.data}
           postNum={postNum}
           commentMode={post?.kind === "t1"}
+          commentsDirect={commentsDirect}
         />
       )}
 
       {/* Click wrappter */}
-      <div className="select-none" onClick={(e) => handleClick(e)}>
-        {/* OG Card */}
-        {/* <h1>{postNum}</h1> */}
+      <div className="select-none" > 
         {post?.kind === "t1" ? (
-          <CommentCard data={post?.data} postNum={postNum} />
+          <CommentCard data={post?.data} postNum={postNum} handleClick={handleClick}/>
         ) : context?.cardStyle === "row1" ? (
           <Row1
             post={post?.data}
@@ -148,6 +154,8 @@ const Post = ({ post, postNum = 0 }) => {
             hideNSFW={hideNSFW}
             forceMute={forceMute}
             postNum={postNum}
+            read={read}
+            handleClick={handleClick}
           />
         ) : context?.cardStyle === "card2" ? (
           <Card2
@@ -156,6 +164,8 @@ const Post = ({ post, postNum = 0 }) => {
             hideNSFW={hideNSFW}
             forceMute={forceMute}
             postNum={postNum}
+            read={read}
+            handleClick={handleClick}
           />
         ) : (
           <Card1
@@ -164,6 +174,8 @@ const Post = ({ post, postNum = 0 }) => {
             hideNSFW={hideNSFW}
             forceMute={forceMute}
             postNum={postNum}
+            read={read}
+            handleClick={handleClick}
           />
         )}
       </div>