+++ src/components/PostOptButton.tsx
@@ -11,6 +11,7 @@ import SaveButton from "./SaveButton";
 import HideButton from "./HideButton";
 import { useMainContext, localRead } from "../MainContext";
 import useFilterSubs from "../hooks/useFilterSubs";
+import { useRead } from "../hooks/useRead";
 
 const MyLink = (props) => {
   let { href, children, ...rest } = props;
@@ -24,32 +25,20 @@ const MyLink = (props) => {
 const PostOptButton = ({ post, postNum, mode = "" }) => {
   const context: any = useMainContext();
   const { addSubFilter, addUserFilter } = useFilterSubs();
-  //check store if read and update context
-  const [read, setRead] = useState(false);
   const filterMenuRef = useRef(null);
-  useEffect(() => {
-    const checkRead = async () => {
-      let read = await localRead.getItem(post?.name);
-      if (read) {
-        context.addReadPost(post?.name);
-        setRead(true);
-      }
-    };
-    checkRead();
-  }, []);
+  const {read} = useRead(post?.name)
 
-  useEffect(() => {
-    context?.readPosts?.[post?.name] ? setRead(true) : setRead(false);
-  }, [context?.readPosts]);
 
   const toggleRead = async () => {
     context?.toggleReadPost(post?.name);
-    setRead((r) => !r);
   };
 
   return (
     <>
-      <Menu as="div" className={" relative font-normal "}>
+      <Menu
+        as="div"
+        className={" relative font-normal "}
+      >
         {({ open }) => (
           <>
             <div
@@ -149,7 +138,6 @@ const PostOptButton = ({ post, postNum, mode = "" }) => {
                     )}
                   </Menu.Item>
                 )}
-               
 
                 {post?.link_flair_text && (
                   <Menu.Item disabled={post?.link_flair_text ? false : true}>
@@ -174,7 +162,7 @@ const PostOptButton = ({ post, postNum, mode = "" }) => {
                     )}
                   </Menu.Item>
                 )}
-                 <Menu.Item>
+                <Menu.Item>
                   {({ active }) => (
                     <div
                       onClick={(e) => {
@@ -265,7 +253,7 @@ const PostOptButton = ({ post, postNum, mode = "" }) => {
                         className="flex flex-row items-center min-w-full px-2 py-1 text-sm hover:cursor-pointer"
                       >
                         <MdOutlineClear className="flex-none w-4 h-4 mr-2 mt-0.5 " />
-                        <h1>{read ? "Mark Unread" : "Mark Read"}</h1>
+                        <h1>{read  ? "Mark Unread" : "Mark Read"}</h1>
                       </button>
                     </div>
                   )}