+++ src/components/cards/Card2.tsx
@@ -11,15 +11,15 @@ import PostOptButton from "../PostOptButton";
 import { GoRepoForked } from "react-icons/go";
 
 //og card
-const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
+const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum, read, handleClick }) => {
   const context: any = useMainContext();
   return (
-    <div>
+    <div onClick={(e) => handleClick(e)}>
       <div
         className={
           (context?.columnOverride == 1 && "") +
           " text-sm bg-lightPost group hover:bg-lightPostHover dark:hover:bg-darkPostHover hover:shadow-2xl transition-colors border hover:cursor-pointer border-gray-300 shadow-md dark:bg-darkBG dark:border-trueGray-700 dark:hover:border-trueGray-500 hover:border-gray-400" +
-          " rounded-lg overflow-clip"
+          " rounded-lg overflow-clip" 
         }
       >
         <div className="">
@@ -33,9 +33,10 @@ const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
               hideNSFW={hideNSFW}
               post={post}
               forceMute={forceMute}
-              allowIFrame={false}
               postMode={false}
               imgFull={false}
+              read={read}
+              card={true}
             />
           </a>
           {true && (
@@ -48,6 +49,7 @@ const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
                       (post?.distinguished == "moderator" || post?.stickied
                         ? " text-lightGreen dark:text-darkGreen "
                         : " ")
+                        + (read && context.dimRead ? " opacity-50" : "")
                     }
                   >{`${post?.title ?? ""}`}</span>
                 </a>
@@ -149,7 +151,7 @@ const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
                   />
                 </div>
                 <div className="flex flex-row items-center gap-2 ml-auto mr-6">
-                  <a href={post?.permalink} onClick={(e) => e.preventDefault()}>
+                  <a href={post?.permalink} onClick={(e) => {e.preventDefault(); e.stopPropagation(); handleClick(e, true);}}>
                     <h1
                       className={
                         "cursor-pointer hover:underline font-semibold " +
@@ -163,7 +165,7 @@ const Card1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
                   </a>
                 </div>
                 <div className="absolute right-3">
-                  <PostOptButton post={post} postNum={postNum} />
+                  <PostOptButton post={post} postNum={postNum}  />
                 </div>
               </div>
             </div>