+++ src/components/cards/Row1.tsx
@@ -20,11 +20,21 @@ import { hideLink } from "../../RedditAPI";
 import HideButton from "../HideButton";
 import PostOptButton from "../PostOptButton";
 import { GoRepoForked } from "react-icons/go";
-const Row1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
+const Row1 = ({
+  post,
+  hasMedia,
+  hideNSFW,
+  forceMute,
+  postNum,
+  read,
+  handleClick,
+}) => {
+  const context: any = useMainContext();
   const [expand, setexpand] = useState(false);
 
   return (
     <div
+      onClick={(e) => handleClick(e)}
       className={
         (postNum === 0 ? " border-t rounded-t-md " : " ") +
         "flex flex-row items-start py-1 pb-2 text-sm bg-lightPost dark:bg-[#212121] dark:hover:bg-darkPostHover group hover:bg-lightPostHover border-l border-r border-transparent  dark:hover:border-trueGray-500 hover:border-gray-500  "
@@ -119,7 +129,8 @@ const Row1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
                     " group-hover:underline font-semibold text-base " +
                     (post?.distinguished == "moderator" || post?.stickied
                       ? " text-lightGreen dark:text-darkGreen "
-                      : " ")
+                      : " ") +
+                    (read && context.dimRead ? " opacity-50" : "")
                   }
                 >{`${post?.title ?? ""}`}</span>
               </a>
@@ -243,7 +254,9 @@ const Row1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
             </button>
 
             <a href={post?.permalink} onClick={(e) => e.preventDefault()}>
-              <button className="flex flex-row items-center px-2 py-1 h-[26px] space-x-1 border border-transparent rounded-md hover:border-lightBorderHighlight dark:hover:border-darkBorderHighlight ">
+              <button
+              onClick={e => {e.stopPropagation(); e.preventDefault(); handleClick(e, true)}}
+              className="flex flex-row items-center px-2 py-1 h-[26px] space-x-1 border border-transparent rounded-md hover:border-lightBorderHighlight dark:hover:border-darkBorderHighlight ">
                 <BiComment className="flex-none w-4 h-4 " />
                 <h1 className="">{`${
                   numToString(post?.num_comments, 1000) ?? "??"
@@ -313,7 +326,6 @@ const Row1 = ({ post, hasMedia, hideNSFW, forceMute, postNum }) => {
                 hideNSFW={hideNSFW}
                 post={post}
                 forceMute={forceMute}
-                allowIFrame={expand}
                 postMode={false}
                 imgFull={false}
               />