+++ src/components/settings/Toggles.tsx
@@ -34,7 +34,13 @@ type ComponentProps = {
     | "showFlairs"
     | "showUserFlairs"
     | "expandedSubPane"
-    | "infiniteLoading";
+    | "infiniteLoading"
+    | "dimRead"
+    | "autoRead"
+    | "disableEmbeds"
+    | "preferEmbeds" 
+    | "embedsEverywhere";
+
   label?: string;
   externalStyles?: string;
   withSubtext?: boolean;
@@ -146,7 +152,7 @@ const Toggles = ({
         !label && setSwitchLabel("Sync Width");
         !subtext &&
           setSwitchSubtext(
-            "Syncs wide UI changes with post width. If wide UI is disabled posts will not automatically display comments to the side."
+            "Syncs Wide UI changes with Post Width. If Post Width is narrow posts will not automatically display comments to the side."
           );
         break;
       case "postWideUI":
@@ -197,13 +203,45 @@ const Toggles = ({
             "Automatically shows the subreddit dropdown pane as expanded instead of collapsed"
           );
         break;
-        case "infiniteLoading":
+      case "infiniteLoading":
         !label && setSwitchLabel("Infinite Loading");
         !subtext &&
           setSwitchSubtext(
             "Switches between infinite loaded or paginated feeds"
           );
         break;
+      case "dimRead":
+        !label && setSwitchLabel("Dim Read");
+        !subtext && setSwitchSubtext("Dim read post titles and text in cards");
+        break;
+      case "autoRead":
+        !label && setSwitchLabel("Auto Read");
+        !subtext &&
+          setSwitchSubtext(
+            "Automatically mark posts as read when their thread is opened"
+          );
+        break;
+        case "disableEmbeds":
+          !label && setSwitchLabel("Disable Embeds");
+          !subtext &&
+            setSwitchSubtext(
+              "Will not load any embeds unless you explicitly switch to embed"
+            );
+          break;
+          case "preferEmbeds":
+            !label && setSwitchLabel("Prefer Embeds");
+            !subtext &&
+              setSwitchSubtext(
+                "Prefer embeds instead of native video. Native video options may not work (autoplay, hoverplay, audio, etc.)"
+              );
+            break;
+            case "embedsEverywhere":
+              !label && setSwitchLabel("Embed Everywhere");
+              !subtext &&
+                setSwitchSubtext(
+                  "By default embeds will only show in single column view or in a post thread. Enable this to show embeds in multi-column mode. Note, this is disabled by default for better performance."
+                );
+              break;
       default:
         break;
     }
@@ -259,9 +297,24 @@ const Toggles = ({
       case "expandedSubPane":
         context.toggleExpandedSubPane();
         break;
-        case "infiniteLoading":
+      case "infiniteLoading":
         context.toggleInfiniteLoading();
         break;
+      case "dimRead":
+        context.toggleDimRead();
+        break;
+      case "autoRead":
+        context.toggleAutoRead();
+        break;
+        case "disableEmbeds":
+          context.toggleDisableEmbeds();
+          break;
+          case "preferEmbeds":
+          context.togglePreferEmbeds();
+          break;
+          case "embedsEverywhere":
+          context.toggleEmbedsEverywhere();
+          break;
       default:
         break;
     }